Move project-index module into shared layer
Relocated src/plugin/src/project-index/index.js to src/shared/project-index/index.js so scripts and tests can load the index without reaching into plugin internals.

Updated dependent imports and docs to reference the shared location.
